toolkit/components/extensions/test/mochitest/test_chrome_ext_downloads_saveAs.html

<!doctype html>
<html>
<head>
  <title>Test downloads.download() saveAs option</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/AddTask.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script src="chrome_head.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

ChromeUtils.import("resource://gre/modules/FileUtils.jsm");

const PROMPTLESS_DOWNLOAD_PREF = "browser.download.useDownloadDir";

// We need to be able to distinguish files downloaded by the file picker from
// files downloaded without it.
let pickerDir;
let defaultDir;

add_task(async function setup() {
  let downloadDir = FileUtils.getDir("TmpD", ["downloads"]);
  pickerDir = downloadDir.clone();
  pickerDir.createUnique(Ci.nsIFile.DIRECTORY_TYPE, FileUtils.PERMS_DIRECTORY);
  info(`Using file picker download directory ${pickerDir.path}`);
  defaultDir = downloadDir.clone();
  defaultDir.createUnique(Ci.nsIFile.DIRECTORY_TYPE, FileUtils.PERMS_DIRECTORY);
  info(`Using default download directory ${defaultDir.path}`);

  isnot(pickerDir.path, defaultDir.path,
        "Should be able to distinguish between files saved with or without the file picker");

  await SpecialPowers.pushPrefEnv({"set": [
    ["browser.download.folderList", 2],
    ["browser.download.dir", defaultDir.path],
  ]});

  SimpleTest.registerCleanupFunction(async () => {
    await SpecialPowers.popPrefEnv();
    pickerDir.remove(true);
    defaultDir.remove(true);
  });
});

add_task(async function test_downloads_saveAs() {
  const pickerFile = pickerDir.clone();
  pickerFile.append("file_download.nonext.txt");

  const defaultFile = defaultDir.clone();
  defaultFile.append("file_download.nonext.txt");

  const {MockFilePicker} = SpecialPowers;
  MockFilePicker.init(window);

  MockFilePicker.showCallback = fp => {
    // On picker 'show' event, choose the filename that was set as the default
    // and append it to the picker's download directory
    let file = pickerDir.clone();

    // Assert that the downloads API configures both default properties.
    is(fp.defaultString, "file_download.nonext.txt", "Got the expected FilePicker defaultString");
    is(fp.defaultExtension, "txt", "Got the expected FilePicker defaultExtension");

    file.append(fp.defaultString);
    MockFilePicker.setFiles([file]);
  };

  function background() {
    const url = URL.createObjectURL(new Blob(["file content"]));
    browser.test.onMessage.addListener(async (filename, saveAs) => {
      try {
        let options = {
          url,
          filename,
        };
        // Only define the saveAs option if the argument was actually set
        if (saveAs !== undefined) {
          options.saveAs = saveAs;
        }
        let id = await browser.downloads.download(options);
        browser.downloads.onChanged.addListener(delta => {
          if (delta.id == id && delta.state.current === "complete") {
            browser.test.sendMessage("done", {ok: true, id});
          }
        });
      } catch ({message}) {
        browser.test.sendMessage("done", {ok: false, message});
      }
    });
    browser.test.sendMessage("ready");
  }

  const manifest = {background, manifest: {permissions: ["downloads"]}};
  const extension = ExtensionTestUtils.loadExtension(manifest);

  await extension.startup();
  await extension.awaitMessage("ready");

  async function testExpectFilePicker(saveAs) {
    ok(!pickerFile.exists(), "the file should have been cleaned up properly previously");

    MockFilePicker.returnValue = MockFilePicker.returnOK;

    extension.sendMessage("file_download.nonext.txt", saveAs);
    let result = await extension.awaitMessage("done");
    ok(result.ok, `downloads.download() works with saveAs=${saveAs}`);

    ok(pickerFile.exists(), "the file exists.");
    is(pickerFile.fileSize, 12, "downloaded file is the correct size");
    pickerFile.remove(false);

    // Test the user canceling the save dialog.
    MockFilePicker.returnValue = MockFilePicker.returnCancel;

    extension.sendMessage("file_download.nonext.txt", saveAs);
    result = await extension.awaitMessage("done");

    ok(!result.ok, "download rejected if the user cancels the dialog");
    is(result.message, "Download canceled by the user", "with the correct message");
    ok(!pickerFile.exists(), "file was not downloaded");
  }

  async function testNoFilePicker(saveAs) {
    ok(!defaultFile.exists(), "the file should have been cleaned up properly previously");

    extension.sendMessage("file_download.nonext.txt", saveAs);
    let result = await extension.awaitMessage("done");
    ok(result.ok, `downloads.download() works with saveAs=${saveAs}`);

    ok(defaultFile.exists(), "the file exists.");
    is(defaultFile.fileSize, 12, "downloaded file is the correct size");
    defaultFile.remove(false);
  }

  info("Testing that saveAs=true uses the file picker as expected");
  let saveAs = true;
  await testExpectFilePicker(saveAs);

  info("Testing that saveAs=false does not use the file picker");
  saveAs = false;
  await testNoFilePicker(saveAs);

  // When saveAs is not set, the behavior should be determined by the Firefox
  // pref that normally determines whether the "Save As" prompt should be
  // displayed.
  info(`Testing that the file picker is used when saveAs is not specified ` +
       `but ${PROMPTLESS_DOWNLOAD_PREF} is disabled`);
  saveAs = undefined;
  await SpecialPowers.pushPrefEnv({"set": [
    [PROMPTLESS_DOWNLOAD_PREF, false],
  ]});
  await testExpectFilePicker(saveAs);

  info(`Testing that the file picker is NOT used when saveAs is not ` +
       `specified but ${PROMPTLESS_DOWNLOAD_PREF} is enabled`);
  await SpecialPowers.popPrefEnv();
  await SpecialPowers.pushPrefEnv({"set": [
    [PROMPTLESS_DOWNLOAD_PREF, true],
  ]});
  await testNoFilePicker(saveAs);

  await extension.unload();
  MockFilePicker.cleanup();
});

</script>

</body>
</html>
